[send-non-sendable] Remove PartitionUtils.h and move its contents into SendNonSendable.
When another pass actually uses PartitionUtils, we can extract it back into a header.
